Add new type inference, under /typeSystemRefresh:1 (#4323)
This PR contains the bulk of the long-awaited changes of the Dafny's new
type system. Most of the changes are to implement a new type inference.
This was necessary, because the previous implementation was very
complicated and difficult to understand and modify. The new type
inference is more straightforward (see, e.g., the description in
https://github.com/dafny-lang/dafny/pull/4134). Unlike the previous type
system, the new one supports traits over datatypes and lays the
groundwork for some other type-system changes we planning in the months
ahead.

The changes in this PR are under the option `--type-system-refresh`
(legacy `/typeSystemRefresh:1`), except for some minor improvements
(e.g., a better error source location for modify statements). (Also, to
use traits over datatypes, it is also necessary to use
`--general-traits` / `/generalTraits:1`.)

While this PR contains the core parts of the new type system, these
changes are not final. However, **it makes sense to get these changes
into the master branch at this time**, to more easily keep other changes
in synch. For this reason, I don't know at what level of detail this PR
ought to be reviewed. Because the changes are isolated to the
`--type-system-refresh` option and because we should do a comprehensive
review of the the system before we switch to making it the default in
Dafny, there will be another chance to review anything that is more
shallowly or leniently reviewed at this time.

From a functionality standpoint, the biggest thing missing from this PR
is inference of subset types. For example, for a class `C`, this PR may
infer the type `C?` rather than the non-null type `C`. In some cases,
this will just make the verifier work harder, but in other cases (e.g.,
when a local variable has such a type and the local variable is modified
inside a loop), the new type inference in this PR may generate an error.
The additional precision that comes from considering subset types will
be added in a separate PR.

One actual _change_ in the type system is worth highlighting: Consider a
`trait Tr` and a `class Cl extends Tr`. In the type system of Dafny
today, `Tr` is assignable to `Cl` without an explicit cast. But in the
new type system, this require an explicit cast. (Note, what I just said
applies to traits, not subset types. So, for example, there is no change
in how an `int` can be assigned to a `nat`.)

## Basic approach

The basic approach (see also the description of
https://github.com/dafny-lang/dafny/pull/4134, whose changes are
included in this PR) is to

* First, infer _pre-types_. These are the parts of types that don't
depend on theorem proving. In other words, they are what you get if you
replace subset types with their base types.
* Then, make the types more precise by considering subset types. (This
step is not included in this PR.)

## Additional notes

In the implementation, some of the resolver's fields had to be made more
accessible (e.g., changing method `FreeVariables` to be `public`).

Error messages generated by the new type inference include the string
`PRE-TYPE:`. This continues to be useful in debugging, but will be
removed by the time we start recommending the new type system.

The PR disables CI coverage tests for the new code, because the test
suite has not yet been changed to use the new type system.

The code contains some TODO's that will be filled in by later PRs. This
is not standard practice, but I don't want to lose track of these TODO's
and the code surrounding them is need to make everything compile and
run.


<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>