[Test Generation] Change the set of coverage criteria that can be targeted (#4326)
This PR changes what coverage criteria a user of test generation may
target.

- Path coverage: Dafny will now attempt complete path coverage of the
(inlined part of the) program as oppose to covering paths through each
method in isolation. As a means of dealing with path explosion, Dafny
will now try generating tests for increasingly longer paths, skipping
over any paths whose subpaths have been found to be infeasible.

- Branch coverage: As an middle ground between path coverage and block
coverage, Dafny can now attempt call-graph-sensitive version of block
coverage (i.e. it tries to cover every block in the program for every
path through the call graph that can lead to that block). I call this
"call-graph" coverage but am not sure this is the best way to call it -
would be grateful for any suggestions.

This PR relies on a [recent change](https://github.com/boogie-org/boogie/pull/753)
in Boogie that fixes a bug in how inlined programs are printed to a file.

By submitting this pull request, I confirm that my contribution
is made under the terms of the MIT license.

---------

Co-authored-by: Aleksandr Fedchin <fedchina@amazon.com>
Co-authored-by: Aaron Tomb <aarontomb@gmail.com>