Fix: Top-level symbols are now all verified. (#4443)
Fixes #4442
In https://github.com/dafny-lang/dafny/pull/4427, a bug was introduced
which prevent multiple top-level verification symbols to be reported,
only the first one would. This adds a test to cover the issue and fixes
it. I verified that the test was failing before the fix, and now the
test is passing. I am not adding a docs/dev/news entry because the bug
was introduced after the 4.2.0 release.

<!-- Is this a user-visible change? Remember to update RELEASE_NOTES.md
-->

<!-- Is this a bug fix? Remember to include a test in Test/git-issues/
-->

<!-- Is this a bug fix for an issue introduced in the latest release?
Mention this in the PR details and ensure a patch release is considered
-->

<!-- Does this PR need tests? Add them to `Test/` or to
`Source/*.Test/…` and run them with `dotnet test` -->

<!-- Are you moving a large amount of code? Read CONTRIBUTING.md to
learn how to do that while maintaining git history -->

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>