Allow cardinality to be treated as `long` in Java (#4484)
Previously, using the cardinality of a collection in a context where it
would be compiled to a `long` in Java led to a type error from the Java
compiler. This change ensures that the necessary cast is inserted.

Fixes #4483

By submitting this pull request, I confirm that my contribution
is made under the terms of the MIT license.